refactor: 에픽 수정 모달 수정
confirm 버튼을 사용하도록 수정
