+++ client/src/components/EpicEntryItem/index.tsx
@@ -4,6 +4,8 @@ import draggableIcon from '@public/icons/draggable.svg';
 import { Modal } from '@/lib/design';
 import { EpicType } from '@/types/epic';
 import EpicEditModal from '../EpicEditModal';
+import { useEpicDispatch } from '@/lib/hooks/useContextHooks';
+import { updateEpicById } from '@/lib/api/epic';
 
 interface EpicEntryItemProps {
   activated: boolean;
@@ -18,6 +20,24 @@ interface EpicEntryItemProps {
 const EpicEntryItem = (props: EpicEntryItemProps) => {
   const [showDraggable, setShowDraggable] = useState(false);
   const [showEditModal, setShowModal] = useState(false);
+  const [value, setValue] = useState(props.epicData.name);
+
+  const dispatchEpic = useEpicDispatch();
+
+  const handleChange = (ev: React.ChangeEvent) => {
+    setValue((ev.target as HTMLInputElement).value);
+  };
+
+  const handleFormSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    const updatedEpic = { ...props.epicData, name: value };
+    updateEpicById(props.epicData.id, updatedEpic);
+    dispatchEpic({
+      type: 'UPDATE_EPIC',
+      epic: updatedEpic,
+    });
+    setShowModal(false);
+  };
 
   return (
     <>
@@ -32,12 +52,18 @@ const EpicEntryItem = (props: EpicEntryItemProps) => {
         <div>{props.epicData.name}</div>
       </S.Container>
       <Modal
-        shouldConfirm={false}
+        shouldConfirm
         title="에픽 제목 수정"
         visible={showEditModal}
         onClose={() => setShowModal(false)}
+        onClickOk={handleFormSubmit}
       >
-        <EpicEditModal setShowModal={setShowModal} epicData={props.epicData} />
+        <EpicEditModal
+          epicData={props.epicData}
+          value={value}
+          handleChange={handleChange}
+          handleFormSubmit={handleFormSubmit}
+        />
       </Modal>
     </>
   );
